# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: HAS SmartLock Service API
  version: 0.1.0
  description: |
    当前仓库已实现接口的 OpenAPI 定义。
    用于导入 Apifox、Postman 等 API 测试平台。
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/device/homeCreate:
    post:
      tags:
        - Home
      summary: 创建家庭
      operationId: homeCreate
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HomeCreateRequest'
            example:
              name: '{{home_name}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/device/homes:
    get:
      tags:
        - Home
      summary: 家庭列表
      operationId: homes
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomesEnvelope'
              example:
                code: 1000
                msg: ok
                data:
                  - id: h_123456789
                    name: My Home
                    location: Shenzhen
                    count: 0
                    role: 1
                    create_time: 1711857600
  /v1/device/bind:
    post:
      tags:
        - Device
      summary: 设备绑定（仅限设备端调用）
      operationId: deviceBind
      description: |
        设备端专属签名接口。
        sign 使用 model_secret 计算，不使用用户端 secret_key 规则。
      parameters:
        - $ref: '#/components/parameters/DeviceModelHeader'
        - $ref: '#/components/parameters/DeviceUUIDHeader'
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceBindRequest'
            example:
              uid: '{{uid}}'
              mac: '{{mac}}'
              zone: '{{zone}}'
              version: '{{firmware_version}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/device/login:
    post:
      tags:
        - Device
      summary: 设备登录（仅限设备端调用）
      operationId: deviceLogin
      description: |
        设备端专属签名接口。
        sign 使用 model_secret 计算，不使用用户端 secret_key 规则。
      parameters:
        - $ref: '#/components/parameters/DeviceModelHeader'
        - $ref: '#/components/parameters/DeviceUUIDHeader'
        - $ref: '#/components/parameters/DeviceUIDHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceLoginRequest'
            example:
              zone: '{{zone}}'
              a: true
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/device/list:
    get:
      tags:
        - Device
      summary: 设备列表
      operationId: deviceList
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
        - in: query
          name: home_id
          required: false
          schema:
            type: string
            example: '{{home_id}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListEnvelope'
              example:
                code: 1000
                msg: ok
                data: []
  /v1/device/newList:
    get:
      tags:
        - Device
      summary: 新设备列表
      operationId: newDeviceList
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewDevicesEnvelope'
              example:
                code: 1000
                msg: ok
                data: []
  /v1/device/upName:
    post:
      tags:
        - Device
      summary: 修改设备名称
      operationId: deviceUpName
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpNameRequest'
            example:
              uuid: '{{uuid}}'
              name: '{{device_name}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/device/homeDevices:
    get:
      tags:
        - Home
      summary: 家庭设备列表
      operationId: homeDevices
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
        - in: query
          name: home_id
          required: true
          schema:
            type: string
            example: '{{home_id}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeDevicesEnvelope'
              example:
                code: 1000
                msg: ok
                data: []
  /v1/device/homeUsers:
    get:
      tags:
        - Home
      summary: 家庭用户列表
      operationId: homeUsers
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
        - in: query
          name: home_id
          required: true
          schema:
            type: string
            example: '{{home_id}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeUsersEnvelope'
              example:
                code: 1000
                msg: ok
                data:
                  - uid: u_123456789
                    username: demo@example.com
                    avatar: avatar/u_123456789
                    role: 1
                    accept: 1
  /v1/device/homeUpdate:
    post:
      tags:
        - Home
      summary: 更新家庭信息
      operationId: homeUpdate
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HomeUpdateRequest'
            example:
              home_id: '{{home_id}}'
              name: '{{home_name}}'
              location: '{{location}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/device/homeDelete:
    delete:
      tags:
        - Home
      summary: 删除家庭
      operationId: homeDelete
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
        - in: query
          name: home_id
          required: true
          schema:
            type: string
            example: '{{home_id}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
...
components:
  parameters:
    AppIDHeader:
      in: header
      name: appid
      required: true
      example: '{{appid}}'
      schema:
        type: string
        default: '{{appid}}'
      description: 应用 ID
    AppVersionHeader:
      in: header
      name: app_version
      required: true
      example: '{{app_version}}'
      schema:
        type: string
        default: '{{app_version}}'
      description: 应用版本号
    PhoneCodeHeader:
      in: header
      name: phone_code
      required: true
      example: '{{phone_code}}'
      schema:
        type: string
        default: '{{phone_code}}'
      description: 手机国家区号
    TimestampHeader:
      in: header
      name: timestamp
      required: true
      example: '{{timestamp}}'
      schema:
        type: string
        default: '{{timestamp}}'
      description: Unix 秒级时间戳
    RequestIDHeader:
      in: header
      name: request_id
      required: true
      example: '{{request_id}}'
      schema:
        type: string
        default: '{{request_id}}'
      description: 请求唯一 ID
    SignHeader:
      in: header
      name: sign
      required: true
      example: '{{sign}}'
      schema:
        type: string
        default: '{{sign}}'
      description: 请求签名
    AccessTokenHeader:
      in: header
      name: access_token
      required: true
      example: '{{access_token}}'
      schema:
        type: string
        default: '{{access_token}}'
      description: 访问令牌
    DeviceModelHeader:
      in: header
      name: model
      required: true
      example: '{{model}}'
      schema:
        type: string
        default: '{{model}}'
      description: 设备型号
    DeviceUUIDHeader:
      in: header
      name: uuid
      required: true
      example: '{{uuid}}'
      schema:
        type: string
        default: '{{uuid}}'
      description: 设备 UUID
    DeviceUIDHeader:
      in: header
      name: uid
      required: true
      example: '{{uid}}'
      schema:
        type: string
        default: '{{uid}}'
      description: 用户 UID
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    EnvelopeBase:
      type: object
      required:
        - code
        - msg
      properties:
        code:
          type: integer
          example: 1000
        msg:
          type: string
          example: ok
...
    NullEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              nullable: true
...
    HomeCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
    HomeUpdateRequest:
      type: object
      required:
        - home_id
        - name
        - location
      properties:
        home_id:
          type: string
        name:
          type: string
        location:
          type: string
    UpNameRequest:
      type: object
      required:
        - uuid
        - name
      properties:
        uuid:
          type: string
        name:
          type: string
    DeviceBindRequest:
      type: object
      required:
        - uid
        - mac
        - zone
        - version
      properties:
        uid:
          type: string
        mac:
          type: string
        zone:
          type: string
        version:
          type: string
    DeviceLoginRequest:
      type: object
      required:
        - zone
        - a
      properties:
        zone:
          type: string
        a:
          type: boolean
...
    HomeItem:
      type: object
      required:
        - id
        - name
        - location
        - count
        - role
        - create_time
      properties:
        id:
          type: string
        name:
          type: string
        location:
          type: string
        count:
          type: integer
        role:
          type: integer
        create_time:
          type: integer
          format: int64
    HomeUser:
      type: object
      required:
        - uid
        - username
        - avatar
        - role
        - accept
      properties:
        uid:
          type: string
        username:
          type: string
        avatar:
          type: string
        role:
          type: integer
        accept:
          type: integer
    DeviceState:
      type: object
      properties:
        desired:
          type: object
          additionalProperties: true
        reported:
          type: object
          additionalProperties: true
    Device:
      type: object
      required:
        - id
        - uuid
        - device_id
        - uid
        - bind_type
        - secret
        - name
        - first_bind_time
        - bind_time
        - State
      properties:
        id:
          type: integer
        uuid:
          type: string
        device_id:
          type: string
        uid:
          type: string
        bind_type:
          type: integer
        secret:
          type: string
        name:
          type: string
        first_bind_time:
          type: integer
          format: int64
        bind_time:
          type: integer
          format: int64
        State:
          $ref: '#/components/schemas/DeviceState'
    NewDevice:
      type: object
      required:
        - uuid
        - device_id
        - uid
        - bind_type
        - secret
        - name
        - bind_status
        - first_bind_time
        - bind_time
        - delete_time
      properties:
        uuid:
          type: string
        device_id:
          type: string
        uid:
          type: string
        bind_type:
          type: integer
        secret:
          type: string
        name:
          type: string
        bind_status:
          type: integer
        first_bind_time:
          type: integer
          format: int64
        bind_time:
          type: integer
          format: int64
        delete_time:
          type: integer
          format: int64
    HomesEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/HomeItem'
    DeviceListEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Device'
    HomeDevicesEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Device'
    NewDevicesEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/NewDevice'
    HomeUsersEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/HomeUser'